Initial Commit
Add to repository the contents of the original
TADS3 package (2014-2019) for ST2 by @VoidPhantom:

    https://github.com/VoidPhantom/sublime-tads

Adopt the MIT License for the new ST4 package.

Add common Git and EditorConfig settings for
a Sublime Text package repository plus Ruby
and AsciiDoc support in view of the upcoming
Rake toolchain and Asciidoctor documentation.